[ecs-ec2-supervisor] Add some useful local tools (#690)
Fixes ES-836.

This PRs extends the Terraform module in the ecs-ec2-supervisor folder with optional automatic creation of a simple ECS EC2 cluster to help with some quick tests.

It still allows users of the module to provide the name of an already existing cluster to be used instead of creating a brand new one.

The additional variables created to control this behavior are documented in the README and in the tfvars example file.