[supervised-collector] Automate release (#707)
# Description

Fixes CX-19577.

This PR introduces release automation for the Supervised Collector
image.

It all works based on the `otel-supervised-collector/CURRENT_VERSION`
file. This file should contain a valid Collector version, which will be
used in many different places:

- By the `install.sh` script to fetch binaries of the
Collector/Supervisor version that we support.
- By the Supervised Collector CI to detect when a release should happen.
If the contents of the file changed, a new release will be made.

# How Has This Been Tested?

- Did some local tests with the updated install script.

Unfortunately more tests for the workflow triggers will only be possible
after merging.

# Checklist:
- [ ] I have updated the relevant Helm chart(s) version(s)
- [ ] I have updated the relevant component changelog(s)
- [ ] This change does not affect any particular component (e.g. it's CI
or docs change)

---------

Signed-off-by: Douglas Camata <159076+douglascamata@users.noreply.github.com>