Add Coralogix exporter and Service pipelines to Advanced config (#724)
# Description

The OTel k8s README sync is now active and the Advanced config section
of the README replaced the current Advanced config in the docs. They
were very similar, except two sections: Coralogix exporter and Service
pipelines.
I added them as they are (except yaml formatting) to the same location
as in the original doc.
Could you please verify the info in the sections and their placement?

# How Has This Been Tested?

mdox

# Checklist:
- [ ] I have updated the relevant Helm chart(s) version(s)
- [ ] I have updated the relevant component changelog(s)
- [ ] This change does not affect any particular component (e.g. it's CI
or docs change)